switch the database to sqlite
