+++ app.go
@@ -1,11 +1,10 @@
 package main
 
 import (
-	Mysql "imgHub/web/Databases"
 	"imgHub/web/Router"
 )
 
 func main() {
 	Router.InitRouter()
-	defer Mysql.DB.Close()
+	// defer Sqlite.DB.Close()
 }
