+++ go.mod
@@ -6,4 +6,6 @@ require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gin-gonic/gin v1.6.3
 	github.com/jinzhu/gorm v1.9.14
+	gorm.io/driver/sqlite v1.1.4 // indirect
+	gorm.io/gorm v1.21.4 // indirect
 )
