+++ upload.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env /Library/Frameworks/Python.framework/Versions/3.9/bin/python3
 import requests
 import sys
