+++ web/Databases/sqlite.go
@@ -1,0 +1,21 @@
+package Sqlite
+
+import (
+	"fmt"
+
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
+)
+
+var DB *gorm.DB
+
+func init() {
+	var err error
+	DB, err = gorm.Open(sqlite.Open("imghub.db"), &gorm.Config{})
+	if err != nil {
+		fmt.Printf("sqlite connect error %v", err)
+	}
+	if DB.Error != nil {
+		fmt.Printf("database error %v", DB.Error)
+	}
+}
