+++ web/Models/viewModel.go
@@ -1,25 +1,24 @@
 package Models
 
 import (
-	"imgHub/web/Databases"
+	Sqlite "imgHub/web/Databases"
 )
 
 type ViewLogs struct {
-	ID int
+	ID    int
 	IMGID string
-	TIME int64
-	IP	string
-	UA	string
+	TIME  int64
+	IP    string
+	UA    string
 }
 
-func init(){
-	Mysql.DB.AutoMigrate(&ViewLogs{})
+func init() {
+	Sqlite.DB.AutoMigrate(&ViewLogs{})
 }
-
 
 // 添加访问日志
-func (this *ViewLogs) Insert() (id int,err error){
-	result := Mysql.DB.Create(&this)
+func (this *ViewLogs) Insert() (id int, err error) {
+	result := Sqlite.DB.Create(&this)
 	id = this.ID
 	if result.Error != nil {
 		err = result.Error
